requirements.txt
apache-airflow==2.9.2
pyspark==3.5.0
requests
psycopg2-binary
python-dotenv
pendulum==2.1.2
apache-airflow-providers-fab==1.2.0
apache-airflow-providers-apache-spark
